% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchevents_operations.R
\name{delete_rule}
\alias{delete_rule}
\title{Deletes the specified rule}
\usage{
delete_rule(Name)
}
\arguments{
\item{Name}{The name of the rule.}
}
\description{
Deletes the specified rule.
}
\details{
Before you can delete the rule, you must remove all targets, using RemoveTargets.

When you delete a rule, incoming events might continue to match to the deleted rule. Allow a short period of time for changes to take effect.
}
\examples{

}
