% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchevents_operations.R
\name{describe_event_bus}
\alias{describe_event_bus}
\title{Displays the external AWS accounts that are permitted to write events to your account using your account's event bus, and the associated policy}
\usage{
describe_event_bus()
}
\description{
Displays the external AWS accounts that are permitted to write events to your account using your account's event bus, and the associated policy. To enable your account to receive events from other accounts, use PutPermission.
}
\examples{

}
