% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchevents_operations.R
\name{describe_rule}
\alias{describe_rule}
\title{Describes the specified rule}
\usage{
describe_rule(
  Name = "string"
)
}
\arguments{
\item{Name}{[required] The name of the rule.}
}
\description{
Describes the specified rule.
}
\details{
DescribeRule does not list the targets of a rule. To see the targets associated with a rule, use ListTargetsByRule.
}
\examples{

}
