% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchevents_operations.R
\name{disable_rule}
\alias{disable_rule}
\title{Disables the specified rule}
\usage{
disable_rule(
  Name = "string"
)
}
\arguments{
\item{Name}{[required] The name of the rule.}
}
\description{
Disables the specified rule. A disabled rule won't match any events, and won't self-trigger if it has a schedule expression.
}
\details{
When you disable a rule, incoming events might continue to match to the disabled rule. Allow a short period of time for changes to take effect.
}
\examples{

}
