% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchevents_operations.R
\name{enable_rule}
\alias{enable_rule}
\title{Enables the specified rule}
\usage{
enable_rule(Name)
}
\arguments{
\item{Name}{The name of the rule.}
}
\description{
Enables the specified rule. If the rule does not exist, the operation fails.
}
\details{
When you enable a rule, incoming events might not immediately start matching to a newly enabled rule. Allow a short period of time for changes to take effect.
}
\examples{

}
