% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchevents_operations.R
\name{list_targets_by_rule}
\alias{list_targets_by_rule}
\title{Lists the targets assigned to the specified rule}
\usage{
list_targets_by_rule(Rule, NextToken = NULL, Limit = NULL)
}
\arguments{
\item{Rule}{The name of the rule.}

\item{NextToken}{The token returned by a previous call to retrieve the next set of results.}

\item{Limit}{The maximum number of results to return.}
}
\description{
Lists the targets assigned to the specified rule.
}
\examples{

}
