% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchevents_package.R
\docType{package}
\name{paws.cloudwatchevents-package}
\alias{paws.cloudwatchevents}
\alias{paws.cloudwatchevents-package}
\title{paws.cloudwatchevents: Amazon CloudWatch Events}
\description{
- Automatically invoke an AWS Lambda function
to update DNS entries when an event notifies you that Amazon EC2
instance enters the running state.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
