% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchevents_operations.R
\name{put_permission}
\alias{put_permission}
\title{Running \code{PutPermission} permits the specified AWS account to put events to your account's default \emph{event bus}}
\usage{
put_permission(Action, Principal, StatementId)
}
\arguments{
\item{Action}{The action that you are enabling the other account to perform. Currently, this must be \code{events:PutEvents}.}

\item{Principal}{The 12-digit AWS account ID that you are permitting to put events to your default event bus. Specify "*" to permit any account to put events to your default event bus.

If you specify "*", avoid creating rules that may match undesirable events. To create more secure rules, make sure that the event pattern for each rule contains an \code{account} field with a specific account ID from which to receive events. Rules with an account field do not match any events sent from other accounts.}

\item{StatementId}{An identifier string for the external account that you are granting permissions to. If you later want to revoke the permission for this external account, specify this \code{StatementId} when you run RemovePermission.}
}
\description{
Running \code{PutPermission} permits the specified AWS account to put events to your account's default \emph{event bus}. CloudWatch Events rules in your account are triggered by these events arriving to your default event bus.
}
\details{
For another account to send events to your account, that external account must have a CloudWatch Events rule with your account's default event bus as a target.

To enable multiple AWS accounts to put events to your default event bus, run \code{PutPermission} once for each of these accounts.

The permission policy on the default event bus cannot exceed 10 KB in size.
}
\examples{

}
