% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchevents_operations.R
\name{remove_targets}
\alias{remove_targets}
\title{Removes the specified targets from the specified rule}
\usage{
remove_targets(Rule, Ids)
}
\arguments{
\item{Rule}{The name of the rule.}

\item{Ids}{The IDs of the targets to remove from the rule.}
}
\description{
Removes the specified targets from the specified rule. When the rule is triggered, those targets are no longer be invoked.
}
\details{
When you remove a target, when the associated rule triggers, removed targets might continue to be invoked. Allow a short period of time for changes to take effect.

This action can partially fail if too many requests are made at the same time. If that happens, \code{FailedEntryCount} is non-zero in the response and each entry in \code{FailedEntries} provides the ID of the failed target and the error code.
}
\examples{

}
