% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchevents_package.R
\docType{package}
\name{paws.cloudwatchevents-package}
\alias{paws.cloudwatchevents}
\alias{paws.cloudwatchevents-package}
\title{paws.cloudwatchevents: Amazon CloudWatch Events}
\description{
Amazon CloudWatch Events helps you to respond to
    state changes in your AWS resources.  When your resources change
    state, they automatically send events into an event stream.  You can
    create rules that match selected events in the stream and route them
    to targets to take action.  You can also use rules to take action on a
    pre-determined schedule.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
