% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchevents_operations.R
\name{put_events}
\alias{put_events}
\title{Sends custom events to Amazon CloudWatch Events so that they can be matched to rules}
\usage{
put_events(
  Entries = list(
    list(
      Time = as.POSIXct("2015-01-01"),
      Source = "string",
      Resources = list(
        "string"
      ),
      DetailType = "string",
      Detail = "string"
    )
  )
)
}
\arguments{
\item{Entries}{[required] The entry that defines an event in your system. You can specify several parameters for the entry such as the source and type of the event, resources associated with the event, and so on.}
}
\description{
Sends custom events to Amazon CloudWatch Events so that they can be matched to rules.
}
\examples{

}
